pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
]

[tool.uv.workspace]
members = [
    "ai_wolf",
]
